requirements.txt
Flask==3.0.3
gunicorn==21.2.0
supabase==2.6.0
google-generativeai==0.7.2
pydantic==2.9.2
pytest==8.3.3
pytest-cov==5.0.0
requests==2.32.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0

